Added a better solution for Eagle_BP
Updated the Eagle_BP with an invisible vision cone. The Eagle_BP is storing actors it sees in an array, but with a major bug.

To do: remove the array for the Eagle_BP in which it checks the actor for being the same actor in next tick. Add a TimeSeen variable in every actor instead.